serial timeout from uint16_t to uint32_t

+++ components/bm1397/serial.c
@@ -65,7 +65,7 @@ int SERIAL_send(uint8_t *data, int len, bool debug)
 /// @param buf buffer to read data into
 /// @param buf number of ms to wait before timing out
 /// @return number of bytes read, or -1 on error
-int16_t SERIAL_rx(uint8_t *buf, uint16_t size, uint16_t timeout_ms)
+int16_t SERIAL_rx(uint8_t *buf, uint16_t size, uint32_t timeout_ms)
 {
     int16_t bytes_read = uart_read_bytes(UART_NUM_1, buf, size, timeout_ms / portTICK_PERIOD_MS);
     // if (bytes_read > 0) {
